refactoring code services and controllers

+++ src/modules/user/user.service.ts
@@ -18,17 +18,22 @@ export class UserService {
     @InjectRepository(Role)
     private readonly rolesRepository: Repository<Role>,
   ) {}
-  async createUser(userRegisterDto: CreateUserDto) {
+
+  createUser(userRegisterDto: CreateUserDto) {
     const user = this.userRepository.create(userRegisterDto);
     return this.userRepository.save(user);
   }
 
-  async findAll() {
-    return await this.userRepository.find();
+  findAll() {
+    const user = this.userRepository.find();
+    if (!user) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+    return user;
   }
 
-  async findOneById(id: string) {
-    const getOneUser = await this.userRepository.findOne({
+  findOneById(id: string) {
+    const getOneUser = this.userRepository.findOne({
       where: { id },
       relations: { permissions: true, roles: true },
     });
@@ -38,22 +43,35 @@ export class UserService {
     return getOneUser;
   }
 
-  async findOneByEmail(email: string) {
-    return await this.userRepository.findOne({ where: { email } });
+  findOneByEmail(email: string) {
+    const user = this.userRepository.findOne({ where: { email } });
+    if (!user) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+    return user;
   }
 
-  async update(id: string, updateUserDto: UpdateUserDto) {
-    return this.userRepository.update(
+  update(id: string, updateUserDto: UpdateUserDto) {
+    const userExists = this.userRepository.findOneBy({ id });
+    if (!userExists) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+    const user = this.userRepository.update(
       { id },
       {
         name: updateUserDto.name,
         email: updateUserDto.email,
       },
     );
+    return user;
   }
 
-  async deleteUser(id: string) {
-    const deleteUser = await this.userRepository.delete({
+  deleteUser(id: string) {
+    const userExists = this.userRepository.findOneBy({ id });
+    if (!userExists) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+    const deleteUser = this.userRepository.delete({
       id,
     });
     return deleteUser;
@@ -69,18 +87,15 @@ export class UserService {
     if (!user) {
       throw new HttpException('User not found', HttpStatus.NOT_FOUND);
     }
-
     const permisionExists = await this.permissionRepository.findBy({
       id: In(userACLRequest.permissions),
     });
     const rolesExists = await this.rolesRepository.findBy({
       id: In(userACLRequest.roles),
     });
-
     user.permissions = permisionExists;
     user.roles = rolesExists;
     user.save();
-
     return user;
   }
 }
